requirements.txt
django
djangorestframework
django-heroku
gunicorn
pyyaml
